requirements.txt
cx-Freeze==6.15.0
cx-Logging==3.1.0
lief==0.13.1
psutil==5.9.5
pypresence==4.2.1
websocket-client==1.5.2
